backend/requirements.txt
flask==3.1.1
flask-cors==6.0.2
firebase-admin==7.4.0
requests==2.33.1
python-dotenv==1.2.2